+++ tests/test_MCAPWeightCalculator.py
@@ -2,7 +2,7 @@ import pypoanal.portfolio_calculators as pc
 import pandas as pd
 import datetime
 import numpy as np
-from hypothesis import given, example
+from hypothesis import given
 from hypothesis import strategies as st
 from pytest import approx
 
@@ -11,10 +11,11 @@ def test_MCAPWeightsCalculator_nans():
     mcap_calc = pc.compute_mcap_weights
     ticker_nshares = pd.Series({'GOOG': 100.0, 'AMZN': 100.0})
     price_history = pd.DataFrame({'GOOG': [100.0, 100.0], 'AMZN': [100.0, np.nan]},
-                                 index=[datetime.date(2010,1,1),
-                                        datetime.date(2010,1,2)])
+                                 index=[datetime.date(2010, 1, 1),
+                                        datetime.date(2010, 1, 2)])
     weights: pd.Series = mcap_calc(ticker_nshares, price_history)
     assert weights.to_dict() == {'GOOG': 0.5, 'AMZN': 0.5}
+
 
 def test_MCAPWeightsCalculator_equal_1():
     mcap_calc = pc.compute_mcap_weights
@@ -36,7 +37,7 @@ def test_MCAPWeightsCalculator_1():
 
 def test_MCAPWeightsCalculator_empy_shares():
     mcap_calc = pc.compute_mcap_weights
-    ticker_nshares = pd.Series()
+    ticker_nshares = pd.Series(dtype=np.float64)
     price_history = pd.DataFrame({'GOOG': [200.0], 'AMZN': [100.0]},
                                  index=[datetime.date(2010, 1, 1)])
     weights: pd.Series = mcap_calc(ticker_nshares, price_history)
@@ -51,7 +52,9 @@ def test_MCAPWeightsCalculator_nan_history():
     weights: pd.Series = mcap_calc(ticker_nshares, price_history)
     assert weights.to_dict() == {}
 
-@given(goog_c=st.integers(min_value=0), amzn_c=st.integers(min_value=0), goog_p=st.floats(min_value=0.0), amzn_p=st.floats(min_value=0.0))
+
+@given(goog_c=st.integers(min_value=0), amzn_c=st.integers(min_value=0), goog_p=st.floats(min_value=0.0),
+       amzn_p=st.floats(min_value=0.0))
 def test_MCAPWeightsCalculator_hypothesis(goog_c, amzn_c, goog_p, amzn_p):
     mcap_calc = pc.compute_mcap_weights
     ticker_nshares = pd.Series({'GOOG': goog_c, 'AMZN': amzn_c})
@@ -59,5 +62,3 @@ def test_MCAPWeightsCalculator_hypothesis(goog_c, amzn_c, goog_p, amzn_p):
                                  index=[datetime.date(2010, 1, 1)])
     weights: pd.Series = mcap_calc(ticker_nshares, price_history)
     assert weights.empty or weights.sum() == approx(1.0)
-
-
